requirements.txt
kafka-python==2.0.2
elasticsearch==8.8.0
Flask==2.3.2
celery==5.3.1